Add scss files and node-sass. Breakup reducers into seperate files
