+++ app/scripts/components/App.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import NavBar from './NavBar.jsx';
 
 const App = (props) => (
-  <main>
+  <main className="outer_container">
     <NavBar />
     { props.children }
   </main>
