+++ app/scripts/reducers/reducer.js
@@ -1,49 +1,12 @@
 import { combineReducers } from 'redux';
+// import { reducer as formReducer } 'redux-form';
 
-const initUserState = {
-  valid: false,
-  zip: null,
-  username: null,
-  password: null,
-};
+import { users } from './reducer_users.js';
+import { items } from './reducer_items.js';
 
-const users = (state = initUserState, action) => {
-  switch (action.type) {
-    case 'LOGIN_VALID':
-      return Object.assign({}, state, {
-        valid: action.valid,
-        zip: action.zip,
-      });
-    case 'ID_USER':
-      return Object.assign({}, state, {
-        username: action.username,
-        password: action.password,
-      });
-    default:
-      return state;
-  }
-};
-
-const initItemState = {
-  items: [],
-};
-
-const items = (state = initItemState, action) => {
-  switch (action.type) {
-    case 'GET_INITIAL_ITEMS':
-      return Object.assign({}, state, {
-        items: action.items,
-      });
-    default:
-      return state;
-  }
-};
-
-const reducer = combineReducers({
+const rootReducer = combineReducers({
   items,
   users,
 });
 
-module.exports = {
-  reducer,
-};
+export default rootReducer;
