+++ app/scripts/reducers/reducer_items.js
@@ -1,0 +1,18 @@
+
+
+const initItemState = {
+  items: [],
+};
+
+const items = (state = initItemState, action) => {
+  switch (action.type) {
+    case 'GET_INITIAL_ITEMS':
+      return Object.assign({}, state, {
+        items: action.items,
+      });
+    default:
+      return state;
+  }
+};
+
+export { items };
