+++ app/scripts/reducers/reducer_users.js
@@ -1,0 +1,27 @@
+
+
+const initUserState = {
+  valid: false,
+  zip: null,
+  username: null,
+  password: null,
+};
+
+const users = (state = initUserState, action) => {
+  switch (action.type) {
+    case 'LOGIN_VALID':
+      return Object.assign({}, state, {
+        valid: action.valid,
+        zip: action.zip,
+      });
+    case 'ID_USER':
+      return Object.assign({}, state, {
+        username: action.username,
+        password: action.password,
+      });
+    default:
+      return state;
+  }
+};
+
+export { users };
