+++ public/css/styles.css
@@ -1,0 +1,120 @@
+/* http://meyerweb.com/eric/tools/css/reset/
+   v2.0 | 20110126
+   License: none (public domain)
+*/
+html, body, div, span, applet, object, iframe,
+h1, h2, h3, h4, h5, h6, p, blockquote, pre,
+a, abbr, acronym, address, big, cite, code,
+del, dfn, em, img, ins, kbd, q, s, samp,
+small, strike, strong, sub, sup, tt, var,
+b, u, i, center,
+dl, dt, dd, ol, ul, li,
+fieldset, form, label, legend,
+table, caption, tbody, tfoot, thead, tr, th, td,
+article, aside, canvas, details, embed,
+figure, figcaption, footer, header, hgroup,
+menu, nav, output, ruby, section, summary,
+time, mark, audio, video {
+  margin: 0;
+  padding: 0;
+  border: 0;
+  font-size: 100%;
+  font: inherit;
+  vertical-align: baseline; }
+
+/* HTML5 display-role reset for older browsers */
+article, aside, details, figcaption, figure,
+footer, header, hgroup, menu, nav, section {
+  display: block; }
+
+body {
+  line-height: 1; }
+
+ol, ul {
+  list-style: none; }
+
+blockquote, q {
+  quotes: none; }
+
+blockquote:before, blockquote:after,
+q:before, q:after {
+  content: '';
+  content: none; }
+
+table {
+  border-collapse: collapse;
+  border-spacing: 0; }
+
+/*
+====================
+Nav Bar header
+====================
+ */
+nav {
+  width: 960px;
+  margin: 0 auto;
+  margin-bottom: 24px;
+  border-bottom: 1px solid #000; }
+
+h1 {
+  font-size: 32px; }
+
+.auth_points {
+  padding: 6px;
+  display: inline-block;
+  text-align: right; }
+
+.app_points {
+  padding: 6px;
+  display: inline-block;
+  text-align: left; }
+
+.nav_title {
+  text-align: center;
+  font-size: 24px; }
+
+.dashboard,
+.create {
+  padding-left: 24px; }
+
+.login,
+.signup {
+  padding-right: 24px; }
+
+/*
+====================
+Main Container
+====================
+ */
+.main_container {
+  width: 960px;
+  margin: 0 auto; }
+
+/*
+====================
+Inputs - Login / Signup
+====================
+ */
+input, textarea {
+  display: block; }
+
+/*
+====================
+Flex Grid
+====================
+ */
+.outer_container {
+  display: flex;
+  max-width: 960px;
+  margin: 0 auto;
+  border: 1px solid black; }
+
+nav {
+  display: flex;
+  max-width: 960px;
+  margin: 0 auto; }
+
+.ListView_container {
+  display: flex;
+  max-width: 960px;
+  margin: 0 auto; }
