+++ public/index.html
@@ -3,8 +3,9 @@
   <head>
     <title>discollect</title>
     <meta charset="utf-8">
-    <link rel="stylesheet" href="reset.css">
+    <!-- <link rel="stylesheet" href="cssreset.css"> -->
     <link rel="stylesheet" href="styles.css">
+    <!-- <link rel="stylesheet" href="css/styles.css"> -->
   </head>
   <body>
     <div id="main">
